feat:  create ITestResultReporter
